% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generatePopulation}
\alias{generatePopulation}
\title{Create a random population}
\usage{
generatePopulation(n = 99, dimension = matrix(rpois(25, lambda = 10), ncol =
  5), continuous.mean = c(5, 10, 7), continuous.var = 2 *
  diag(length(continuous.mean)), categorical.proportion = c(1/6, 1/3, 1/2))
}
\arguments{
\item{n}{the size of the population}

\item{dimension}{A matrix with the proportion for the size of the rectangles}

\item{continuous.mean}{A vector of population means}

\item{continuous.var}{A variance covariance matrix of the population means}

\item{categorical.proportion}{A vector with proportions}
}
\description{
Create a random population
}

