% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotPopulation}
\alias{plotPopulation}
\title{Display the population}
\usage{
plotPopulation(population, rows, cols)
}
\arguments{
\item{population}{A data.frame with the population as the output of \code{\link{generatePopulation}}}

\item{rows}{the number of rows}

\item{cols}{the number of columns}
}
\description{
Display the population
}

